lambda/requirements.txt
aws-requests-auth==0.4.1
certifi==2018.4.16
elasticsearch==6.2.0
elasticsearch-curator==5.5.4
